refactor admin tutorials page
